# packages/contracts/openapi/novel-factory.v1.yaml
openapi: 3.1.0
info:
  title: Novel Factory API
  version: 1.0.0
servers:
  - url: /v1
paths:
...
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /workspaces/{workspaceId}/home:
    get:
      operationId: getWorkspaceHome
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace home aggregate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
...
  /writing-runs/{writingRunId}:
    get:
      operationId: getWritingRun
      parameters:
        - in: path
          name: writingRunId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Writing run detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
        '404':
          description: Writing run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /writing-runs/{writingRunId}/accept-chapter:
    post:
      operationId: acceptWritingChapter
      parameters:
        - in: path
          name: writingRunId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Writing chapter accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
        '404':
          description: Writing run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '409':
          description: Writing run acceptance dependencies incomplete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /quality-reports/{qualityReportId}:
    get:
      operationId: getQualityReport
      parameters:
        - in: path
          name: qualityReportId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Quality report detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
        '404':
          description: Quality report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /feedback-records:
    get:
      operationId: listFeedbackRecords
      parameters:
        - in: query
          name: targetType
          required: false
          schema:
            type: string
        - in: query
          name: targetId
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Feedback records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
...
components:
  schemas:
...
    ApiSuccessEnvelope:
      $ref: ../schemas/api-success-envelope.schema.json
    ApiErrorEnvelope:
      $ref: ../schemas/api-error-envelope.schema.json
...